syscalls: Correct __floatunsisf and __floatunsidf
int, not unsigned int and not float/double. Correct sign extension

+++ syscalls/softfloat/implementation/__floatunsisf_helper.java
@@ -1,8 +1,8 @@
-public static float __floatunsisf_helper(int i)
+public static int __floatunsisf_helper(int i)
 {
 	long v = (long)i;
 	
 	v &= 0x00000000ffffffffl;
 
-	return Float.floatToIntBits((int)v);
+	return Float.floatToIntBits((float)v);
 }
